chore: move plugin inside src
