+++ eslint.config.mjs
@@ -33,6 +33,6 @@ export default defineConfig([
     },
   },
   {
-    ignores: ['node_modules/', 'lib/', 'plugin/build/'],
+    ignores: ['node_modules/', 'lib/'],
   },
 ]);
