+++ package.json
@@ -50,7 +50,7 @@
     "typecheck": "tsc",
     "lint": "eslint \"**/*.{js,ts,tsx}\"",
     "clean": "git clean -fxd",
-    "prepare": "./scripts/set-version.sh && bob build && tsc --build plugin",
+    "prepare": "./scripts/set-version.sh && bob build",
     "start": "rm yarn.lock && yarn && yarn prepare && husky"
   },
   "devDependencies": {
