--- plugin/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "expo-module-scripts/tsconfig.plugin",
-  "compilerOptions": {
-    "outDir": "build",
-    "rootDir": "src",
-    "declaration": true
-  },
-  "include": ["./src"],
-  "exclude": ["**/__mocks__/*", "**/__tests__/*"]
-}
