+++ src/plugin/__tests__/_mock_MainActivity.ts
