+++ src/plugin/__tests__/_mock_emptyAppDelegate.ts
