+++ src/plugin/__tests__/_mock_malformedAppDelegate.ts
