+++ src/plugin/__tests__/setApplicationContinueUserActivitySwift.test.ts
