+++ src/plugin/__tests__/setApplicationOpenUrlSwift.test.ts
