+++ src/plugin/__tests__/setKotlinMainActivity.test.ts
