+++ src/plugin/setAppTheme.ts
