+++ src/plugin/setApplicationContinueUserActivity.ts
