+++ src/plugin/setApplicationContinueUserActivitySwift.ts
