+++ src/plugin/setApplicationOpenUrl.ts
