+++ src/plugin/setApplicationOpenUrlSwift.ts
