+++ src/plugin/setEntitlements.ts
