+++ src/plugin/setJavaMainActivity.ts
