+++ src/plugin/setKotlinMainActivity.ts
