+++ src/plugin/utils/findClassEndIndex.ts
